// src/database/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  USER: 'USER',
  GUEST: 'GUEST'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ProjectStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]


export const ProjectCategory = {
  FULLSTACK: 'FULLSTACK',
  BACKEND: 'BACKEND',
  FRONTEND: 'FRONTEND',
  MOBILE: 'MOBILE',
  DEVOPS: 'DEVOPS',
  AI_ML: 'AI_ML',
  SYSTEMS: 'SYSTEMS',
  OPEN_SOURCE: 'OPEN_SOURCE'
} as const

export type ProjectCategory = (typeof ProjectCategory)[keyof typeof ProjectCategory]


export const SkillCategory = {
  LANGUAGES: 'LANGUAGES',
  BACKEND_FRAMEWORKS: 'BACKEND_FRAMEWORKS',
  FRONTEND_FRAMEWORKS: 'FRONTEND_FRAMEWORKS',
  DATABASES: 'DATABASES',
  DEVOPS_CLOUD: 'DEVOPS_CLOUD',
  TOOLS_ARCHITECTURE: 'TOOLS_ARCHITECTURE',
  TESTING_SECURITY: 'TESTING_SECURITY'
} as const

export type SkillCategory = (typeof SkillCategory)[keyof typeof SkillCategory]


export const EmploymentType = {
  FULL_TIME: 'FULL_TIME',
  PART_TIME: 'PART_TIME',
  CONTRACT: 'CONTRACT',
  FREELANCE: 'FREELANCE',
  INTERNSHIP: 'INTERNSHIP',
  SELF_EMPLOYED: 'SELF_EMPLOYED'
} as const

export type EmploymentType = (typeof EmploymentType)[keyof typeof EmploymentType]


export const ContactStatus = {
  UNREAD: 'UNREAD',
  READ: 'READ',
  REPLIED: 'REPLIED',
  ARCHIVED: 'ARCHIVED',
  SPAM: 'SPAM'
} as const

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus]


export const MetricType = {
  PAGE_VIEW: 'PAGE_VIEW',
  RESUME_DOWNLOAD: 'RESUME_DOWNLOAD',
  PROJECT_CLICK: 'PROJECT_CLICK',
  SOCIAL_CLICK: 'SOCIAL_CLICK'
} as const

export type MetricType = (typeof MetricType)[keyof typeof MetricType]